Project_4_Safaricom_DataWarehouse/requirements.txt
boto3>=1.26.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=0.19.0
pytest>=7.0.0
